[IMP] Rework cases v2 api to use more BPs & less error handling
Error handling has been migrated to the root API V2 blueprint, that will handle them automatically. :)
